requirements.txt
paho-mqtt==2.1.0
flet==0.28.3